pyproject.toml
[project]
name = "athena"
version = "0.1.0"
description = "Athena is a simple OpenAI based voice assistant."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chime>=0.7.0",
    "google-api-python-client>=2.153.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "numpy>=2.1.3",
    "openai>=1.54.4",
    "pvporcupine>=3.0.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scipy>=1.14.1",
    "setuptools>=75.4.0",
    "sounddevice>=0.5.1",
    "soundfile>=0.12.1",
    "toml>=0.10.2",
    "webrtcvad>=2.0.10",
]
